{{cookiecutter.directory_name}}/requirements.txt
flask
flask-sqlalchemy
pymysql
uwsgi
confight